final_pro/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
idna==2.8
mistune==0.8.4
mysqlclient==1.4.4
Pillow==6.1.0
requests==2.22.0

